refactor: streamline Tokenizer conversion implementations and add tokenizer_helpers module
